rename tests for better categorization
