+++ .env
@@ -160,9 +160,8 @@ _APP_VCS_GITEA_BROWSER_ENDPOINT=http://localhost:9515
 _APP_VCS_GITEA_CLIENT_ID=
 _APP_VCS_GITEA_CLIENT_SECRET=
 _APP_VCS_GITEA_WEBHOOK_SECRET=gitea-webhook-secret
-_APP_VCS_WEBHOOK_URL=http://appwrite/v1
 _APP_VCS_GITLAB_CLIENT_ID=
 _APP_VCS_GITLAB_CLIENT_SECRET=
 _APP_VCS_GITLAB_WEBHOOK_SECRET=
-_APP_VCS_WEBHOOK_URL=
+_APP_VCS_WEBHOOK_URL=http://appwrite/v1
 _APP_WEBHOOK_MAX_FAILED_ATTEMPTS=10
